+++ packages/contracts/contracts/Interfaces/ICommunityIssuance.sol
@@ -8,7 +8,7 @@ interface ICommunityIssuance {
     
     event ZEROTokenAddressSet(address _zeroTokenAddress);
     event CommunityPotAddressSet(address _communityPotAddress);
-    event TotalZEROIssuedUpdated(uint _totalZEROIssued);
+    event TotalZEROIssuedUpdated(uint256 _totalZEROIssued);
 
     // --- Functions ---
 
@@ -32,7 +32,7 @@ interface ICommunityIssuance {
     /// @notice sends ZERO tokens to given account
     /// @param _account account to receive the tokens
     /// @param _ZEROamount amount of tokens to transfer
-    function sendZERO(address _account, uint _ZEROamount) external;
+    function sendZERO(address _account, uint256 _ZEROamount) external;
 
     /// @notice This function allows depositing tokens into the community pot for the community to use.
     ///         and configures the deploymentTime if it's the first time this function is called.
@@ -40,5 +40,5 @@ interface ICommunityIssuance {
     /// @param _ZEROamount The amount of ZERO to deposit into the community pot.
     /// @dev   Even if ZeroToken is a trusted ERC20 token contract, it is still important to ensure that
     ///        non reentrancy is possible (maybe due to an upgrade)
-    function receiveZero(address _account, uint _ZEROamount) external;
+    function receiveZero(address _account, uint256 _ZEROamount) external;
 }
